Refactor some task list methods to import them from other modules
